Run a new version of runrunriprate.  This version calls the function combinefreezetimes2.m which is a variation on numbericgroupcombine that aims to add in the freezing times of rats that is calculated by the estpos variable.  This is a little complicated by the fact that if other time filters, such as for velocity, are added to the runriprate filter, then there is another constraint put on the time that is totaled.  For example, I first ran it with a linear velocity < .1 setting in runriprate timefilter.
In runrunriprate, there is now a variable calculated called freezedays, which is the mean/sem of the freezing across animals for all the days (except Dickison day 4 = why not included???).

runrunriprate has a new section to calculate the cumulative freezing across the epoch.  This is not very interesting for most control tracks with little freezing.  With the shock tracks, I did not see the pattern I was expecting with most of the freezing happening at the beginning.  Some days follow this pattern, but other days do not.